+++ Coditech.Project/API/Comman/Coditech.API.Client/Endpoint/Organisation/OrganisationCentreEndpoint.cs
@@ -15,7 +15,7 @@ namespace Coditech.API.Endpoint
         public string CreateOrganisationAsync() =>
             $"{CoditechAdminSettings.CoditechOrganisationApiRootUri}/OrganisationCentreMaster/CreateOrganisationCentre";
 
-        public string GetOrganisationAsync(short organisationCentreId) =>
+        public string GetOrganisationAsync(int organisationCentreId) =>
             $"{CoditechAdminSettings.CoditechOrganisationApiRootUri}/OrganisationCentreMaster/GetOrganisationCentre?organisationCentreMasterId={organisationCentreId}";
        
         public string UpdateOrganisationAsync() =>
@@ -23,52 +23,52 @@ namespace Coditech.API.Endpoint
 
         public string DeleteOrganisationAsync() =>
                   $"{CoditechAdminSettings.CoditechOrganisationApiRootUri}/OrganisationCentreMaster/DeleteOrganisationCentre";
-        public string GetPrintingFormatAsync(short organisationCentreId) =>
+        public string GetPrintingFormatAsync(int organisationCentreId) =>
            $"{CoditechAdminSettings.CoditechOrganisationApiRootUri}/OrganisationCentreMaster/GetPrintingFormat?organisationCentreMasterId={organisationCentreId}";
 
         public string UpdatePrintingFormatAsync() =>
                $"{CoditechAdminSettings.CoditechOrganisationApiRootUri}/OrganisationCentreMaster/UpdatePrintingFormat";
 
-        public string GetCentrewiseGSTSetupAsync(short organisationCentreId) =>
+        public string GetCentrewiseGSTSetupAsync(int organisationCentreId) =>
            $"{CoditechAdminSettings.CoditechOrganisationApiRootUri}/OrganisationCentreMaster/GetCentrewiseGSTSetup?organisationCentreMasterId={organisationCentreId}";
 
         public string UpdateCentrewiseGSTSetupAsync() =>
                $"{CoditechAdminSettings.CoditechOrganisationApiRootUri}/OrganisationCentreMaster/UpdateCentrewiseGSTSetup";
 
-        public string GetCentrewiseSmtpSetupAsync(short organisationCentreId) =>
+        public string GetCentrewiseSmtpSetupAsync(int organisationCentreId) =>
            $"{CoditechAdminSettings.CoditechOrganisationApiRootUri}/OrganisationCentreMaster/GetCentrewiseSmtpSetup?organisationCentreMasterId={organisationCentreId}";
 
         public string UpdateCentrewiseSmtpSetupAsync() =>
                $"{CoditechAdminSettings.CoditechOrganisationApiRootUri}/OrganisationCentreMaster/UpdateCentrewiseSmtpSetup";
-        public string GetCentrewiseSmsSetupAsync(short organisationCentreId, byte generalSmsProviderId) =>
+        public string GetCentrewiseSmsSetupAsync(int organisationCentreId, byte generalSmsProviderId) =>
           $"{CoditechAdminSettings.CoditechOrganisationApiRootUri}/OrganisationCentreMaster/GetCentrewiseSmsSetup?organisationCentreMasterId={organisationCentreId}&generalSmsProviderId={generalSmsProviderId}";
 
         public string UpdateCentrewiseSmsSetupAsync() =>
                $"{CoditechAdminSettings.CoditechOrganisationApiRootUri}/OrganisationCentreMaster/UpdateCentrewiseSmsSetup";
 
-        public string GetCentrewiseWhatsAppSetupAsync(short organisationCentreId, byte generalWhatsAppProviderId) =>
+        public string GetCentrewiseWhatsAppSetupAsync(int organisationCentreId, byte generalWhatsAppProviderId) =>
           $"{CoditechAdminSettings.CoditechOrganisationApiRootUri}/OrganisationCentreMaster/GetCentrewiseWhatsAppSetup?organisationCentreMasterId={organisationCentreId}&generalWhatsAppProviderId={generalWhatsAppProviderId}";
 
         public string UpdateCentrewiseWhatsAppSetupAsync() =>
                $"{CoditechAdminSettings.CoditechOrganisationApiRootUri}/OrganisationCentreMaster/UpdateCentrewiseWhatsAppSetup";
 
-        public string GetCentrewiseEmailTemplateSetupAsync(short organisationCentreId, string emailTemplateCode, string templateType) =>
+        public string GetCentrewiseEmailTemplateSetupAsync(int organisationCentreId, string emailTemplateCode, string templateType) =>
           $"{CoditechAdminSettings.CoditechOrganisationApiRootUri}/OrganisationCentreMaster/GetCentrewiseEmailTemplateSetup?organisationCentreMasterId={organisationCentreId}&emailTemplateCode={emailTemplateCode}&templateType={templateType}";
 
         public string UpdateCentrewiseEmailTemplateSetupAsync() =>
                $"{CoditechAdminSettings.CoditechOrganisationApiRootUri}/OrganisationCentreMaster/UpdateCentrewiseEmailTemplateSetup";
-        public string GetCentrewiseUserNameAsync(short organisationCentreId,short organisationCentrewiseUserNameRegistrationId) =>
+        public string GetCentrewiseUserNameAsync(int organisationCentreId,int organisationCentrewiseUserNameRegistrationId) =>
          $"{CoditechAdminSettings.CoditechOrganisationApiRootUri}/OrganisationCentreMaster/GetCentrewiseUserName?organisationCentreMasterId={organisationCentreId}&organisationCentrewiseUserNameRegistrationId={organisationCentrewiseUserNameRegistrationId}";
         public string UpdateCentrewiseUserNameAsync() =>
               $"{CoditechAdminSettings.CoditechOrganisationApiRootUri}/OrganisationCentreMaster/UpdateCentrewiseUserName";
 
-        public string GetCentrewiseSMSTemplateSetupAsync(short organisationCentreId, string emailTemplateCode) =>
+        public string GetCentrewiseSMSTemplateSetupAsync(int organisationCentreId, string emailTemplateCode) =>
           $"{CoditechAdminSettings.CoditechOrganisationApiRootUri}/OrganisationCentreMaster/GetCentrewiseSMSTemplateSetup?organisationCentreMasterId={organisationCentreId}&emailTemplateCode={emailTemplateCode}";
 
         public string UpdateCentrewiseSMSTemplateSetupAsync() =>
                $"{CoditechAdminSettings.CoditechOrganisationApiRootUri}/OrganisationCentreMaster/UpdateCentrewiseSMSTemplateSetup";
 
-        public string GetCentrewiseWhatsAppTemplateSetupAsync(short organisationCentreId, string emailTemplateCode) =>
+        public string GetCentrewiseWhatsAppTemplateSetupAsync(int organisationCentreId, string emailTemplateCode) =>
           $"{CoditechAdminSettings.CoditechOrganisationApiRootUri}/OrganisationCentreMaster/GetCentrewiseWhatsAppTemplateSetup?organisationCentreMasterId={organisationCentreId}&emailTemplateCode={emailTemplateCode}";
 
         public string UpdateCentrewiseWhatsAppTemplateSetupAsync() =>
