+++ Coditech.Project/API/Comman/Coditech.API.Data/DataModel/HMS/HospitalDoctorAllocatedRoom.cs
@@ -7,7 +7,7 @@ namespace Coditech.API.Data
         [Key]
         public int HospitalDoctorAllocatedOPDRoomId { get; set; }
         public int HospitalDoctorId { get; set; }
-        public short OrganisationCentrewiseBuildingRoomId { get; set; }
+        public int OrganisationCentrewiseBuildingRoomId { get; set; }
         public Nullable<long> CreatedBy { get; set; }
         public Nullable<System.DateTime> CreatedDate { get; set; }
         public Nullable<long> ModifiedBy { get; set; }
